Updated to build in .net 4

+++ src/sharpHDF/Helpers/AttributeHelper.cs
@@ -29,9 +29,9 @@ namespace sharpHDF.Library.Helpers
             ulong n = 0;
 
             AbstractHdf5Object obj =_attributes.ParentObject;
-
+            
             int id = H5A.iterate(obj.Id.Value, H5.index_t.NAME, H5.iter_order_t.NATIVE, ref n,
-                delegate(int _id, IntPtr _namePtr, ref H5A.info_t _ainfo, IntPtr _data)
+                delegate(long _id, IntPtr _namePtr, ref H5A.info_t _ainfo, IntPtr _data)
                 {
                     string attributeName = Marshal.PtrToStringAnsi(_namePtr);
 
@@ -110,8 +110,8 @@ namespace sharpHDF.Library.Helpers
         /// <returns></returns>
         public static Hdf5Attribute GetStringAttribute(Hdf5Identifier _objectId, string _title)
         {
-            int attributeId = 0;
-            int typeId = 0;
+            long attributeId = 0;
+            long typeId = 0;
 
             attributeId = H5A.open(_objectId.Value, _title);
             typeId = H5A.get_type(attributeId);
