+++ src/sharpHDF/Helpers/IdHelper.cs
@@ -3,17 +3,14 @@
  * See license in repo for more information                                  *
  * https://github.com/sharpHDF/sharpHDF                                      *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
-using System;
 using sharpHDF.Library.Structs;
 
-namespace sharpHDF.Library.Helpers
-{
-    internal static class IdHelper
-    {
+namespace sharpHDF.Library.Helpers {
+    internal static class IdHelper {
 #if HDF5_VER1_10
         public static Hdf5Identifier ToId(this Int64 _value)
 #else
-        public static Hdf5Identifier ToId(this Int32 _value)
+        public static Hdf5Identifier ToId(this long _value)
 #endif
         {
             var id = new Hdf5Identifier(_value);
